Added the handler for swiftmailer: * added the suggested dependencies * created the notification * created the handler * added specs for both of them * tests should be fine
I think, on the long run, the EmailNotificationInterface should be replaced in order to make it more general, as it only accept one recipient address (BAD!) -- that was the case of our first email handler, Emailvision.

Right now the SwiftMailer notification just accept a Swift_Message that you can customize as much as you want. Look forward to clean things a little bit for v2 (for example, NotificationInterface#getMessage() should return a 'mixed', not a string).